Improve models and general cleanup.
This commit improves the our blog models in various ways:
- We specify ``parent_page_types`` and ``subpage_types`` for consistent
behaviour in the admin backend.
- We rearrange the model class definitions so that the index page
precedes the ``BlogPage`` model class for more intuitive reading.
- We get rid of the custom context extension for ``BlogIndexPage`` as
wagtail provides all this custom functionality as part of the page
objects manager. As a consequence templates have been adjusted
accordingly to use the build in manager methods.

Also:
- Fixed wrong template variables ``Page`` has been renamed to ``page``
and works now as expected.
- Index page uses its title as html title as well.
- Fixed wrong html comment pre/suffixes
- Removed obsolete explicit hardcoded template path.